added open vins and open cv to install vio script

+++ scripts/install_vio.sh
@@ -1,12 +1,31 @@
 #!/bin/bash
 
-
-# Clone of IMU message merge
-cd ~/git
-git clone git@mrs.felk.cvut.cz:visual-localization/realsense_to_imu.git
-ln -s ~/git/realsense_to_imu ~/workspace/src
-
 default=n
+
+#Realsense IMU synchro
+while true; do
+  [[ -t 0 ]] && { read -t 5 -n 2 -p $'\e[1;32mInstall RealSense IMU message merge? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
+  response=`echo $resp | sed -r 's/(.*)$/\1=/'`
+
+  if [[ $response =~ ^(y|Y)=$ ]]
+  then
+
+    cd ~/git
+    git clone git@mrs.felk.cvut.cz:visual-localization/realsense_to_imu.git
+    ln -s ~/git/realsense_to_imu ~/workspace/src
+
+    echo "Compiling RealSense IMU message merge"
+    cd ~/workspace
+    catkin build
+
+    break
+  elif [[ $response =~ ^(n|N)=$ ]]
+  then
+    break
+  else
+    echo " What? \"$resp\" is not a correct answer. Try y+Enter."
+  fi
+done
 
 #MSCKF
 while true; do
@@ -26,8 +45,7 @@ while true; do
     ln -s ~/git/msckf ~/workspace/src
     ln -s ~/git/msckf_republisher ~/workspace/src
 
-    echo "Compiling first batch of workspaces"
-    # Build some part otherwise it wont build all
+    echo "Compiling MSCKF"
     cd ~/workspace
     catkin build
 
@@ -165,6 +183,58 @@ while true; do
     ln -s ~/git/vins-mono ~/workspace/src
 
     echo "Building VINS-Mono"
+    cd ~/workspace
+    catkin build
+
+    break
+  elif [[ $response =~ ^(n|N)=$ ]]
+  then
+    break
+  else
+    echo " What? \"$resp\" is not a correct answer. Try y+Enter."
+  fi
+done
+
+#Open CV for Open VINS
+while true; do
+  [[ -t 0 ]] && { read -t 5 -n 2 -p $'\e[1;32mInstall Open CV 3.4.6 (for Open VINS)? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
+  response=`echo $resp | sed -r 's/(.*)$/\1=/'`
+
+  if [[ $response =~ ^(y|Y)=$ ]]
+  then
+
+    cd ~/git
+    git clone --branch 3.4.6 https://github.com/opencv/opencv/
+    git clone --branch 3.4.6 https://github.com/opencv/opencv_contrib/
+    mkdir opencv/build/
+    cd opencv/build/
+    cmake -DOPENCV_EXTRA_MODULES_PATH=../../opencv_contrib/modules ..
+    make -j$(nproc)
+    sudo make install
+
+    break
+  elif [[ $response =~ ^(n|N)=$ ]]
+  then
+    break
+  else
+    echo " What? \"$resp\" is not a correct answer. Try y+Enter."
+  fi
+done
+
+#Open VINS
+while true; do
+  [[ -t 0 ]] && { read -t 5 -n 2 -p $'\e[1;32mInstall Open VINS? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
+  response=`echo $resp | sed -r 's/(.*)$/\1=/'`
+
+  if [[ $response =~ ^(y|Y)=$ ]]
+  then
+    cd ~/git
+
+    # Clone of VINS-Mono repo
+    git clone git@mrs.felk.cvut.cz:visual-localization/open_vins.git
+    ln -s ~/git/open_vins ~/workspace/src
+
+    echo "Building Open VINS"
     cd ~/workspace
     catkin build
